added quick scan

+++ Main/delete.php
@@ -2,13 +2,21 @@
 include 'db.php';
 
 $id = $_GET['id'];
-$sql = "DELETE FROM users WHERE id=$id";
 
-if ($conn->query($sql) === TRUE) {
-    header('Location: register.php');
+
+if (isset($id) && is_numeric($id)) {
+    $sql = "DELETE FROM users WHERE id=$id";
+
+    if ($conn->query($sql) === TRUE) {
+        header('Location: users.php'); 
+        exit(); 
+    } else {
+        echo "Error deleting record: " . $conn->error;
+    }
 } else {
-    echo "Error deleting record: " . $conn->error;
+    echo "Invalid ID parameter."; 
 }
 
 $conn->close();
 ?>
+<a href="users.php">Back</a>
